// pacific-sdk/src/code/types.ts
export type FigmaUser = {
  id: string;
  photoURL: string;
  name: string;
};

// Describes the NODE_ENV environment
export type Environment = 'development' | 'production';

// Higher order generic for describing transactional return values
export type Transaction<Value> = [Value, true] | [null, false];

// Describes a prompt configuration
export type PromptConfiguration = {
  temperature: number;
  threshold: number;
  timeout: number;
};
// ...
export type EmanationSDKScreen = {
  meta: {
    theme: EmanationSDKTheme;
  };
  name: string; // E.g. "Profile"
  description: string; // E.g. "The profile screen"
  root: WorldviewNode;
};

export type EmanationSDKApp = {
  meta: {};

  name: string; // E.g. "Finance app"
  description: string; // E.g. "A personal finance app"
  screens: EmanationSDKScreen[];
};

export type EmanationSDKEpoch = {
  name: string; // E.g. "" + Date.now()
  apps: EmanationSDKApp[];
};